Added development container
This makes development more comfortable because the build intermediates
(e.g. the generated bases classes for services) are available and VSCode
can make full use of the CMake integration.

The Dockerfile for the development container has various improvements
over the original one (which is still kept for CI and might be adjusted
later):

- Non-root user added which will automatically be updated to the current
  user's UID and GID. It has sudo permissions in case we want to install
  some further packages temporarily.

- The "apt-get" command is run with cache mounts as recommended here:
    https://docs.docker.com/build/cache/optimize/#use-cache-mounts
  This speeds up the build if only a few new packages are added, because
  previously downloaded packages can be used from the cache. Obviously,
  this won't work if the cache is cleared.

- "--no-install-recommends" is used to avoid installing more than
  100 unnecessary packages. A few requirements are explicitly listed to
  ensure that the build still works. Otherwise e.g. include like
  "math.h" and "<cstdio>" are missing and there is a message
  "plugin needed to handle lto object".

The "build-binary.sh" script is adjusted to run within the development
container. It now takes an optional argument to specify the preset (it
defaults to "Release" as before).